Move AdminProfile to component
